Update README.md
The main goal of this project is to propose a solution that normalize given text information, remove unecessary data and keep only important information in the output. The given dataset is a csv file consisting of writer’s name and other information and the normalization version of it. For this task my suggestion is
the utilization of a Seq2Seq (Sequence-to-Sequence) model is a type of neural network architecture designed to convert one sequence of data into another such the goal of this project. Seq2Seq has been widely used before for tasks like
Language translation and Text summarization.